Use imports + custom loader for examples
