+++ docs/_shared/svgIcons/GithubIcon.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import SvgIcon from '@material-ui/core/SvgIcon';
+import { SvgIcon } from '@material-ui/core';
 
 const GitHub = props => (
   <SvgIcon {...props}>
