+++ docs/pages/api/datepicker/index.mdx
@@ -2,6 +2,12 @@ import Example from '_shared/Example'
 import PageMeta from '_shared/PageMeta'
 import PropTypesTable from '_shared/PropTypesTable'
 
+import * as CustomDay from './CustomDay.example'
+import * as BasicDatePicker from './BasicDatePicker.example'
+import * as ViewsDatePicker from './ViewsDatePicker.example'
+import * as InlineDatePicker from './InlineDatePicker.example'
+import * as KeyboardDatePicker from './KeyboardDatePicker.example'
+
 <PageMeta component="Datepicker" />
 
 ## Date picker
@@ -12,16 +18,13 @@ color and type weight.
 
 #### Basic usage
 
-<Example
-  title="Basic usage"
-  sourceFile="api/datepicker/BasicDatePicker.example.jsx"
-/>
+<Example source={BasicDatePicker} />
 
 #### Keyboard Input
 
 It is recommended to use keyboard input with mask for better desktop experience. More information about [mask](https://github.com/insin/inputmask-core#pattern)
 
-<Example sourceFile="api/datepicker/KeyboardDatePicker.example.jsx" />
+<Example source={KeyboardDatePicker} />
 
 #### Different views
 
@@ -31,16 +34,16 @@ Thus you can easily create
 - Month picker
 - Month + year picker
 
-<Example sourceFile="api/datepicker/ViewsDatePicker.example.jsx" />
+<Example source={ViewsDatePicker} />
 
 #### Inline mode
 
-<Example sourceFile="api/datepicker/InlineDatePicker.example.jsx" />
+<Example source={InlineDatePicker} />
 
 #### Customization
 
 It is allowed to hardly customize displaying of dates. Thus you can add badges or fully change displaying of day.
 
-<Example sourceFile="api/datepicker/CustomDay.example.jsx" />
+<Example source={CustomDay} />
 
 <PropTypesTable src="DatePicker/DatePickerModal.tsx" />
