+++ docs/pages/api/timepicker/index.mdx
@@ -2,6 +2,11 @@ import Example from '_shared/Example'
 import PageMeta from '_shared/PageMeta'
 import PropTypesTable from '_shared/PropTypesTable'
 
+import * as BasicTimePicker from './BasicTimePicker.example'
+import * as InlineTimePicker from './InlineTimePicker.example'
+import * as SecondsTimePicker from './SecondsTimePicker.example'
+import * as KeyboardTimePicker from './KeyboardTimePicker.example'
+
 <PageMeta component="Timepicker" />
 
 
@@ -13,19 +18,19 @@ The selected time is indicated by the filled circle at the end of the clock hand
 #### Basic usage
 A time picker should adjusts to a user’s preferred time setting, i.e. the 12-hour or 24-hour format.
 
-<Example sourceFile="api/timepicker/BasicTimePicker.example.jsx" />
+<Example source={BasicTimePicker} />
 
 #### Keyboard input
 
-<Example sourceFile="api/timepicker/KeyboardTimePicker.example.jsx" />
+<Example source={KeyboardTimePicker} />
 
 #### Inline mode
 
-<Example sourceFile="api/timepicker/InlineTimePicker.example.jsx" />
+<Example source={InlineTimePicker} />
 
 #### Seconds
 Seconds input can be used for selection of precise time point
 
-<Example sourceFile="api/timepicker/SecondsTimePicker.example.jsx" />
+<Example source={SecondsTimePicker} />
 
 <PropTypesTable src="TimePicker/TimePickerModal.tsx" />
