+++ docs/pages/getting-started/usage.mdx
@@ -1,10 +1,7 @@
 import PageMeta from '_shared/PageMeta';
-import NoSsr from '@material-ui/core/NoSsr';
+import { NoSsr } from '@material-ui/core';
 
-<PageMeta
-  title="Usage"
-  description="Simple usage example of material-ui-pickers"
-/>
+<PageMeta title="Usage" description="Simple usage example of material-ui-pickers" />
 
 ## Usage
