+++ docs/pages/guides/controlling-programmatically.mdx
@@ -1,5 +1,6 @@
 import Example from '_shared/Example'
 import PageMeta from '_shared/PageMeta'
+import * as ControllingExample from './ControllingProgrammatically.example'
 
 <PageMeta title="Controlling programmatically" />
 
@@ -7,4 +8,4 @@ import PageMeta from '_shared/PageMeta'
 
 Any picker can be controlled by `ref` property. That can be used for openning any picker from the code. See an example below:
 
-<Example sourceFile="guides/ControllingProgrammatically.example.jsx" />
+<Example source={ControllingExample} />
