+++ docs/pages/guides/css-overrides.mdx
@@ -1,5 +1,6 @@
 import Example from '_shared/Example';
 import PageMeta from '_shared/PageMeta'
+import * as CssOverrides from './CssOverrides.example'
 
 <PageMeta title="CSS Overrides" />
 
@@ -17,7 +18,7 @@ overrides.
 #### Override example
 You can find the override component name and class in the generated classnames for pickers components.
 
-<Example sourceFile="guides/CssOverrides.example.jsx" />
+<Example source={CssOverrides} />
 
 
 #### Typescript
