+++ docs/pages/guides/form-integration.mdx
@@ -1,5 +1,6 @@
 import Example from '_shared/Example';
-import PageMeta from '_shared/PageMeta'
+import PageMeta from '_shared/PageMeta';
+import * as FormikExample from './Formik.example';
 
 <PageMeta title="Form integration" />
 
@@ -12,6 +13,7 @@ the box, so it may be tricky to integrate pickers to the form.
 Here are some examples!
 
 #### Formik Integration
+
 Here is an example of how to integrate material-ui-pickers with Formik
 
-<Example sourceFile="guides/Formik.example.jsx" />
+<Example source={FormikExample} />
