+++ docs/pages/guides/static-components.mdx
@@ -1,5 +1,6 @@
 import Example from '_shared/Example'
 import PageMeta from '_shared/PageMeta'
+import * as StaticComponents from './StaticComponents.example'
 
 <PageMeta title="Static componentes" />
 
@@ -15,4 +16,4 @@ chosen date and submitting state logic.
 #### Static components example
 Please make sure that your imports are consistent
 
-<Example sourceFile="guides/StaticComponents.example.jsx" />
+<Example source={StaticComponents} />
