+++ docs/pages/localization/date-fns.mdx
@@ -1,5 +1,6 @@
 import Example from '_shared/Example';
 import PageMeta from '_shared/PageMeta'
+import * as DateFnsExample from './Date-fns.example.jsx'
 
 <PageMeta title="Date-fns localizataion - material-ui-pickers" />
 
@@ -10,5 +11,5 @@ Date-fns localization simply performs by passing date-fns locale object to the M
 #### Example
 *Note*: Any picker in component tree will be rerendered automatically on locale change
 
-<Example sourceFile="localization/Date-fns.example.jsx" />
+<Example source={DateFnsExample} />
