+++ docs/pages/localization/moment.mdx
@@ -1,5 +1,6 @@
 import Example from '_shared/Example';
 import PageMeta from '_shared/PageMeta'
+import * as MomentExample from './Moment.example'
 
 <PageMeta title="Moment localization - material-ui-pickers" />
 
@@ -14,5 +15,5 @@ locale change.
 #### Example
 *Note*: Any picker in component tree will be rerendered automatically on locale change
 
-<Example sourceFile="localization/Moment.example.jsx" />
+<Example source={MomentExample} />
