+++ docs/pages/localization/persian.mdx
@@ -1,5 +1,6 @@
 import Example from '_shared/Example'
 import PageMeta from '_shared/PageMeta'
+import * as PersianExample from './Persian.example.jsx'
 
 <PageMeta title="Jalaali localization - material-ui-pickers" />
 
@@ -17,4 +18,4 @@ npm install @date-io/jalaali
 #### Example
 You can use the examples below. It is recommended that you change the font.
 
-<Example sourceFile="localization/Persian.example.jsx" />
+<Example source={PersianExample} />
